44 Chain Syntax Grammar
Cleaning up a little the classes
Debugged the issue with last group after an alternative
Finished the first version of the grammar parser
Need to debug in the UI to import the grammar, the keyword and then the custom rules
